backend/requirements.txt
scipy
numpy
Cython
scikit-image
opencv-python-headless
Pillow
matplotlib
imgaug>=0.4
albumentations>0.5
graphviz
tqdm
pyyaml
easydict
torch>=1.4.0
torchvision>=0.5.0
tensorboard
future
cffi
ninja
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.45
